[PATRO23-227] Refactor my misstakes and cleanup user
