+++ src/api/app/User/CreatingUser/CreateUser.cs
@@ -1,4 +1,7 @@
+using Intive.Patronage2023.Api.Keycloak;
+using Intive.Patronage2023.Modules.Budget.Application.Budget;
 using Intive.Patronage2023.Shared.Abstractions.Commands;
+using Newtonsoft.Json;
 
 namespace Intive.Patronage2023.Api.User.CreatingUser;
 
@@ -11,3 +14,54 @@ namespace Intive.Patronage2023.Api.User.CreatingUser;
 /// <param name="Password">Password which user provides.</param>
 /// <param name="Email">Email which user provides.</param>
 public record CreateUser(string Avatar, string FirstName, string LastName, string Password, string Email) : ICommand;
+
+/// <summary>
+/// Sign up request handler.
+/// </summary>
+public class CreateUserCommandHandler : ICommandHandler<CreateUser>
+{
+	private readonly KeycloakService keycloakService;
+
+	/// <summary>
+	/// Initializes a new instance of the <see cref="CreateUserCommandHandler"/> class.
+	/// </summary>
+	/// <param name="keycloakService">KeycloakService.</param>
+	public CreateUserCommandHandler(KeycloakService keycloakService) => this.keycloakService = keycloakService;
+
+	/// <summary>
+	/// Handle sign up operation in keycloak api.
+	/// </summary>
+	/// <param name="command">Command.</param>
+	/// <param name="cancellationToken">A cancellation token that can be used to cancel the command.</param>
+	/// <returns>HttpResponseMessage.</returns>
+	public async Task Handle(CreateUser command, CancellationToken cancellationToken)
+	{
+		var response = await this.keycloakService.GetClientToken(cancellationToken);
+
+		if (!response.IsSuccessStatusCode)
+		{
+			throw new AppException("One or more error occured while trying to create user.");
+		}
+
+		string responseContent = await response.Content.ReadAsStringAsync(cancellationToken);
+
+		if (string.IsNullOrEmpty(responseContent))
+		{
+			throw new AppException("One or more error occured while trying to create user.");
+		}
+
+		Token? token = JsonConvert.DeserializeObject<Token>(responseContent);
+
+		if (token == null || token?.AccessToken == null)
+		{
+			throw new AppException("One or more error occured while trying to create user.");
+		}
+
+		response = await this.keycloakService.AddUser(command, token.AccessToken, cancellationToken);
+
+		if (!response.IsSuccessStatusCode)
+		{
+			throw new AppException("One or more error occured while trying to create user.");
+		}
+	}
+}
