+++ src/api/app/User/SignIn/SignInUser.cs
@@ -1,7 +1,7 @@
 using Intive.Patronage2023.Api.Keycloak;
 using Intive.Patronage2023.Shared.Abstractions.Queries;
 
-namespace Intive.Patronage2023.Api.User.Queries;
+namespace Intive.Patronage2023.Api.User.SignIn;
 
 /// <summary>
 /// SignIn user query.
